Add Chapter 10 exercises
I didn't implement the separate process writes that were mentioned in
the chapter, but the rest is done.

This includes a bunch of executable scripts in profile_cache to profile
a variety of conditions in erlang 17.0 and 18.0+. The ratio mentioned in
the book (concurrent requests have a much lower throughput than
synchronous requests) is off for higher versions of erlang. Is this due
to the dirty scheduler changes? I'll need to look into it more.