Add Chapter 12 exercises
This chapter was more focused on playing around in the terminal.

As a note, my implementation uses mnesia instead of file folders as the
database. As a result, I couldn't really test whether the database
replication worked on my single machine, but I think it probably did.
I'll test it with VMs or something later on.