Add Chapter 13 exercises
There will be further commits as I do some related work on the todo app
(upgrading elixir versions, moving to distillery, deploying with
edeliver), but this concludes the book exercises.

I put `rel` in the `.gitignore` file because it was polluting the commit
with log/other generated files. I'm not sure whether this is a
best-practice, so it may change later on.