Configure edeliver with distillery
This update also adds the "/" path for convenience–it lists the
appropriate curl commands to interact with the application.

My goal with this exercise was to learn how to use edeliver for easy
build/deployment. It's clear that a configuration management tool is
also needed, so my next commit will likely be integrating Chef or
Terraform.

Edeliver also requires a `vm.args` and `todo.config` to be written to
each server in the project folder. Setting the appropriate env variables
(BUILD_HOST and PRODUCTION_HOSTS) is also required for use.

Note: I couldn't get the LINK_VM_ARGS setting to work, so I just put
everything in the default location.